Rework compiler warning changes

+++ subprojects/ce-login/celogin/src/JsmnUtils.cpp
@@ -80,7 +80,8 @@ void JsmnUtils::PrintJsmnToken(const JsmnState& jsmnStateParm,
     }
     else
     {
-        printf("ERROR: Unable to print invalid JsmnToken %lu\n", tokenIdxParm);
+        printf("ERROR: Unable to print invalid JsmnToken %llu\n",
+               (unsigned long long)tokenIdxParm);
     }
 }
